VQA-init/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.3.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0  
bitarray==2.9.2
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm  # Replace with PyPI version
contourpy==1.1.1
cycler==0.12.1
debugpy  # Replace with PyPI version
decorator  # Replace with PyPI version
dnspython==2.6.1
docopt==0.6.2
einops==0.3.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing  # Replace with PyPI version
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==4.29.0
gradio_client==0.16.1
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib-metadata  # Replace with PyPI version
importlib_resources==6.4.0
ipdb==0.13.4
ipykernel  # Replace with PyPI version
ipython  # Replace with PyPI version
jedi  # Replace with PyPI version
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==1.5.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client  # Replace with PyPI version
jupyter_core  # Replace with PyPI version
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline  # Replace with PyPI version
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
munch==2.5.0
nest-asyncio  # Replace with PyPI version
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.9.0.80
orjson==3.10.3
packaging==21.3
pandas==1.1.5
parso  # Replace with PyPI version
pexpect  # Replace with PyPI version
pickleshare  # Replace with PyPI version
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs  # Replace with PyPI version
prompt-toolkit  # Replace with PyPI version
protobuf==5.26.1
psutil  # Replace with PyPI version
ptyprocess  # Replace with PyPI version
pure-eval  # Replace with PyPI version
py-cpuinfo==9.0.0
pyarrow==2.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments  # Replace with PyPI version
pyparsing==3.1.2
python-dateutil  # Replace with PyPI version
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq  # Replace with PyPI version
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.3
sacred==0.8.2
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six  # Replace with PyPI version
smmap==5.0.1
sniffio==1.3.1
stack-data  # Replace with PyPI version
starlette==0.37.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm @ git+https://github.com/rwightman/pytorch-image-models.git@59b3d86c1d69fe85ccb5bbdb2f1711eadae6e4a7
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado  # Replace with PyPI version
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
-e git+https://github.com/dandelin/ViLT.git@762fd3975c180db6fc88f577cf39549983fa373a#egg=vilt
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
xlrd==1.2.0
yarl==1.9.4
